test: move store-sync to workspace vitest
