+++ packages/store-sync/src/sqlite/sqliteTableToSql.test.ts
@@ -16,7 +16,7 @@ describe("sqliteTableToSql", () => {
 
     expect(sql).toMatchInlineSnapshot(
       // eslint-disable-next-line max-len
-      '"create table if not exists \\"some table\\" (\\"x\\" integer not null, \\"y\\" integer not null, \\"name\\" text default \'\' not null, \\"block_number\\" blob default \'1000\' not null, constraint \\"some table__primaryKey\\" primary key (\\"x\\", \\"y\\"))"',
+      `"create table if not exists "some table" ("x" integer not null, "y" integer not null, "name" text default '' not null, "block_number" blob default '1000' not null, constraint "some table__primaryKey" primary key ("x", "y"))"`,
     );
   });
 
@@ -28,9 +28,7 @@ describe("sqliteTableToSql", () => {
 
     const sql = sqliteTableToSql(table);
 
-    expect(sql).toMatchInlineSnapshot(
-      '"create table if not exists \\"some table\\" (\\"name\\" text default \'\' not null)"',
-    );
+    expect(sql).toMatchInlineSnapshot(`"create table if not exists "some table" ("name" text default '' not null)"`);
   });
 
   it("should generate correct SQL when keys != column names", async () => {
@@ -42,7 +40,7 @@ describe("sqliteTableToSql", () => {
     const sql = sqliteTableToSql(table);
 
     expect(sql).toMatchInlineSnapshot(
-      '"create table if not exists \\"some table\\" (\\"snake_case\\" text default \'\' not null, constraint \\"some table__primaryKey\\" primary key (\\"snake_case\\"))"',
+      `"create table if not exists "some table" ("snake_case" text default '' not null, constraint "some table__primaryKey" primary key ("snake_case"))"`,
     );
   });
 });
