+++ packages/store-sync/src/zustand/createStorageAdapter.test.ts
@@ -2,9 +2,9 @@ import { beforeAll, describe, expect, it } from "vitest";
 import { storeEventsAbi } from "@latticexyz/store";
 import { createStorageAdapter } from "./createStorageAdapter";
 import { createStore } from "./createStore";
-import { config, deployMockGame } from "../../test/mockGame";
+import { config, deployMockGame } from "../../../../test-setup/mockGame";
 import { fetchAndStoreLogs } from "../fetchAndStoreLogs";
-import { testClient } from "../../test/common";
+import { testClient } from "../../../../test-setup/common";
 import { getBlockNumber } from "viem/actions";
 import { Address } from "viem";
