--- packages/store-sync/vitestSetup.ts
@@ -1,1 +1,0 @@
-export { setup, cleanup as teardown } from "@ark/attest";
